add content type permissions to the permissions route
