+++ packages/core/admin/server/controllers/__tests__/api-token.test.js
@@ -292,57 +292,4 @@ describe('API Token Controller', () => {
       expect(send).toHaveBeenCalled();
     });
   });
-
-  describe('API tokens permissions layout', () => {
-    const layout = [
-      {
-        label: 'address',
-        subjectId: 'api::address',
-        controllers: [
-          {
-            label: 'address',
-            actions: ['api::address.address.find', 'api::address.address.findOne'],
-          },
-        ],
-      },
-      {
-        label: 'category',
-        subjectId: 'api::category',
-        controllers: [
-          {
-            label: 'category',
-            actions: [
-              'api::category.category.find',
-              'api::category.category.findOne',
-              'api::category.category.create',
-              'api::category.category.update',
-              'api::category.category.delete',
-              'api::category.category.createLocalization',
-            ],
-          },
-        ],
-      },
-    ];
-
-    test('return API tokens layout successfully', async () => {
-      const getApiTokenLayout = jest.fn().mockResolvedValue(layout);
-      const send = jest.fn();
-      const ctx = createContext({}, { send });
-
-      global.strapi = {
-        admin: {
-          services: {
-            'api-token': {
-              getApiTokenLayout,
-            },
-          },
-        },
-      };
-
-      await apiTokenController.getLayout(ctx);
-
-      expect(getApiTokenLayout).toHaveBeenCalled();
-      expect(send).toHaveBeenCalledWith({ data: layout });
-    });
-  });
 });
