+++ packages/core/admin/server/controllers/permission.js
@@ -44,4 +44,10 @@ module.exports = {
       },
     };
   },
+
+  async getContentApiPermissions(ctx) {
+    const actionsMap = await strapi.contentAPI.permissions.getActionsMap();
+
+    ctx.send({ data: actionsMap });
+  },
 };
