+++ packages/core/admin/server/routes/permissions.js
@@ -17,4 +17,12 @@ module.exports = [
       policies: ['admin::isAuthenticatedAdmin'],
     },
   },
+  {
+    method: 'GET',
+    path: '/content-api-permissions',
+    handler: 'permission.getContentApiPermissions',
+    config: {
+      policies: ['admin::isAuthenticatedAdmin'],
+    },
+  },
 ];
